Refined Animations for TripleB including Skinning Necklace Correctly.
I had to start over again from before what I did on friday so I've forgotten / lost a lot of those changes. So it's possible the animations are less to everyones liking as before. I'm currently too exhausted to do much more though, so like a lot of the other animations, they may need touchups.

Also minor issue with Animation Events on the 'shoot' animation but it should be easy fix.