Moved location of end point for points and changed the time required for each grade.
I think the score  system needs to be changed. most probably around the time of sprint 6. Currently it's purely time based, meaning the player can rush past all the enemies and get an S score really easily. There needs to be an incentive to engage in combat.

I believe me and Patrycja agree it should be score based, as in, you get score for killing enemies, and lose score for taking too much time. Rank should be assigned via this score.